% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covm.R
\name{crt_covm3}
\alias{crt_covm3}
\title{Create the covariance matrix for each component of the normal mixture.}
\usage{
crt_covm3(sds, rhos)
}
\arguments{
\item{sds}{The standard deviations for each of the 3 normal components. sd should be a 2-by-3 matrix, each row are standard deviations for the same components.}

\item{rhos}{The Pearson correlation between expression data sources for each component. Rho should be a vector with a length of 3. According to the idr3c model, the first element should be 0 under most conditions.}
}
\description{
This function helps to create the 3 covariance matrix (2-by-2) for each component of the normal mixture.
}
\examples{
sds = cbind(c(1,1),c(1,1),c(1,1))
rhos = c(0,0.84,0.84)
covms = crt_covm3(sds=sds,rhos=rhos)
print(covms)
}

