% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idr3c.R
\name{IDR.3component}
\alias{IDR.3component}
\title{Differential expression analysis by integrating data from multiple sources}
\usage{
IDR.3component(x, para = list(p = c(0.4, 0.3, 0.3), mu = cbind(c(0, 0), c(3,
  3), c(-3, -3)), sd = cbind(c(1, 1), c(1, 1), c(1, 1)), rho = c(0, 0.84,
  0.84)), f_ctrl = list(max.t.inner = 10000, max.t.outer = 10000, b.em =
  1e-06, b.outer = 0.001), disp = list(z.rankplot = FALSE, z.rankplot.cut =
  0.2, outer.trace = TRUE, inner.trace = FALSE))
}
\arguments{
\item{x}{A matrix or data frame of log-fold change of normalized gene expression data, i.e., of non-negative values. The rows correspond to observations (e.g., expression level for a gene), the columns correspond to data sources (labs/platforms etc.). Note that technical replicates (i.e., several sequencing ruins/lanes from the same sample) should be took the average.}

\item{para}{A list that contains the parameter of the normal mixture where the pseudo data were generated.
\itemize{
 \item{$p} {The size of each components (proportion of non/up/down regulated genes)}
 \item{$mu} {Means for each normal component. para$mu should be a 2-by-3 matrix, each row are means for the same components.}
 \item{$sd} {Standard deviations for each normal component. para$sd should be a 2-by-3 matrix, each row are standard deviations for the same components.}
 \item{$rho} {Pearson correlation for each normal component.}
}}

\item{f_ctrl}{A list that contains the parameter controls pseudo-EM procedure. The default \code{f_ctrl} usually works well. 
\itemize{
\item{$max.t.inner} {The maximum iterations numbers of the inner EM algorithm}
\item{$max.t.outer} {The maximum number of times that re-generating pseudo data}
\item{$b.em} {Stop criterion for inner EM algorithm.}
\item{$b.outer} {Stop criterion for outer iteration.}
}}

\item{disp}{A list control the display of the trace of parameter estimation.
\itemize{
 \item{$z.rankplot} {Logical. Should the plot of pseudo data be displayed when updated?}
 \item{$z.rankplot.cut} {The proportion of samples with low idr that will be marked on the plot.}
 \item{$outer.trace} {Logical. Should the trace of the parameter estimation be displayed?}
 \item{$inner.trace} {Logical. Should the trace of the parameter estimation for the inner EM be displayed?}
}}
}
\value{
\code{IDR.3component} produces a list that contain idr value, parameter estimation and other necessary components.
\itemize{
 \item{$para} {The final estimation of the normal mixture}
 \item{$loglik.trace.in} {Trace of log-likelihood of EM produce}
 \item{$loglik.trace.ot} {Trace of log-likelihood of outer iterations}
 \item{$idr} {the posterior probability that a gene is in the non-DEGs group}
 \item{$IDR} {computed irreproducibility rate}
}
}
\description{
Mapping genes expression data from different sources to a normal mixture. Then performing model based cluster and computing irreproducibility rate (idr) of being differential expressed for each genes.
}
\details{
Please refer \bold{Lyu, Y., & Li, Q. (2016). A semi-parametric statistical model for integrating gene expression profiles across different platforms. BMC Bioinformatics, 17(1), 51.}
}
\examples{
x = expr
plot(x[,1],x[,2],xlim=c(-8,8),ylim=c(-8,8),cex = .4);
idr.out = IDR.3component(x = x)
}

